+++ modules/features/settings/src/test/java/au/com/shiftyjelly/pocketcasts/settings/viewmodel/AdvancedSettingsViewModelTest.kt
@@ -39,6 +39,7 @@ class AdvancedSettingsViewModelTest {
         whenever(settings.syncOnMeteredNetwork()).thenReturn(false)
         whenever(settings.backgroundRefreshPodcasts).thenReturn(UserSetting.Mock(true, mock()))
         whenever(settings.prioritizeSeekAccuracy).thenReturn(UserSetting.Mock(false, mock()))
+        whenever(settings.cacheEntirePlayingEpisode).thenReturn(UserSetting.Mock(false, mock()))
         viewModel = AdvancedSettingsViewModel(
             settings,
             analyticsTracker,
@@ -50,5 +51,6 @@ class AdvancedSettingsViewModelTest {
     fun `verify settings methods initialize the viewModel state correctly`() {
         TestCase.assertEquals(viewModel.state.value.backgroundSyncOnMeteredState.isChecked, false)
         TestCase.assertEquals(viewModel.state.value.prioritizeSeekAccuracyState.isChecked, false)
+        TestCase.assertEquals(viewModel.state.value.cacheEntirePlayingEpisodeState.isChecked, false)
     }
 }